### 2022-10-26 #4885 @editorialbot
[PRE REVIEW]: AirTrafficSim: An open-source web-based air traffic simulation platform.
**Submitting author:** @kyfrankie (<a href="http://orcid.org/0000-0002-4905-5411">Ka Yiu Hui</a>)
**Repository:** https://github.com/HKUST-OCTAD-LAB/AirTrafficSim
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.0
