+++ paper.bib
@@ -83,7 +83,7 @@
     number = {730},
     pages = {1999-2049},
     keywords = {climate reanalysis, Copernicus Climate Change Service, data assimilation, ERA5, historical observations},
-    doi = {https://doi.org/10.1002/qj.3803},
+    doi = {10.1002/qj.3803},
     year = {2020}
 }
 
@@ -99,14 +99,16 @@
 }
 
 @article{openap,
-    title={OpenAP: An open-source aircraft performance model for air transportation studies and simulations},
-    author={Sun, Junzi and Hoekstra, Jacco M and Ellerbroek, Joost},
-    journal={Aerospace},
-    volume={7},
-    number={8},
-    pages={104},
-    year={2020},
-    publisher={Multidisciplinary Digital Publishing Institute}
+    AUTHOR = {Sun, Junzi and Hoekstra, Jacco M. and Ellerbroek, Joost},
+    TITLE = {OpenAP: An Open-Source Aircraft Performance Model for Air Transportation Studies and Simulations},
+    JOURNAL = {Aerospace},
+    VOLUME = {7},
+    YEAR = {2020},
+    NUMBER = {8},
+    ARTICLE-NUMBER = {104},
+    URL = {https://www.mdpi.com/2226-4310/7/8/104},
+    ISSN = {2226-4310},
+    DOI = {10.3390/aerospace7080104}
 }
 
 @inproceedings{hkia,
