### 2022-11-08 #4885 @AoifeHughes
Hello @xoolive, @torressa, and @rafalkucharskipk - we're looking for one more reviewer for this project. Would any of you have the time and like to be involved - you can see the article proof [here](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04885/joss.04885/10.21105.joss.04885.pdf) and the repository is located here: https://github.com/HKUST-OCTAD-LAB/AirTrafficSim

Details about reviewing can be found in https://joss.readthedocs.io/en/latest/review_criteria.html, and I'm happy to answer any questions
