### 2022-11-08 #4916 @editorialbot
[REVIEW]: AirTrafficSim: An open-source web-based air traffic simulation platform.
**Submitting author:** @kyfrankie (<a href="http://orcid.org/0000-0002-4905-5411">Ka Yiu Hui</a>)
**Repository:** https://github.com/HKUST-OCTAD-LAB/AirTrafficSim
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.1.0
