+++ airtrafficsim/server/data.py
@@ -95,46 +95,47 @@ class Data:
             JSON CZML file of ERA5 wind data image
         """
         # TODO: Improve data loading to avoid repetitive loading
-        data = xr.open_dataset(Path(__file__).parent.parent.parent.joinpath('data/weather/era5/',file.split('-', 1)[0]+'/multilevel.nc')).sel(level=900, time=datetime.fromisoformat(time), method="pad")
-        data = data.where((((data.latitude >= lat1) & (data.latitude <= lat2)) & ((data.longitude >= (long1+360.0) % 360.0) & (data.longitude <= (long2+360.0) % 360.0))), drop=True)
-        fig = Figure(figsize=(long2-long1, lat2-lat1), facecolor='none', dpi=500)
-        ax = fig.add_axes([0, 0, 1, 1], projection=ccrs.PlateCarree(), frameon=False)
-        ax.set_extent([long1, long2, lat1, lat2], crs=ccrs.PlateCarree())
-        # ax.set_global()
-        # ax.coastlines()  
-        ax.barbs(data.longitude.values, data.latitude.values, data.u.values, data.v.values, 
-                flagcolor='grey',
-                # sizes=dict(emptybarb=0.25, spacing=0.2, height=0.5), linewidth=0.95, length=5, 
-                transform=ccrs.PlateCarree())
-        buf = BytesIO()
-        fig.savefig(buf, format="png")
-        uri = "data:image/png;base64," + base64.b64encode(buf.getbuffer()).decode("ascii")
-        return [
-            {
-                "id": "document",
-                "name": "Weather",
-                "version": "1.0",
-            },
-            {
-                "id": "Weather",
-                "rectangle": {
-                    "coordinates": {
-                        "wsenDegrees": [long1, lat1, long2, lat2],
-                    },
-                    "height": 0,
-                    "fill": True,
-                    "material": {
-                        "image": {
-                            "image": { "uri": uri },
-                            "color": {
-                                "rgba": [255, 255, 255, 128],
+        if Path(__file__).parent.parent.parent.joinpath('data/weather/era5/',file.split('-', 1)[0]).is_dir():
+            data = xr.open_dataset(Path(__file__).parent.parent.parent.joinpath('data/weather/era5/',file.split('-', 1)[0]+'/multilevel.nc')).sel(level=900, time=datetime.fromisoformat(time), method="pad")
+            data = data.where((((data.latitude >= lat1) & (data.latitude <= lat2)) & ((data.longitude >= (long1+360.0) % 360.0) & (data.longitude <= (long2+360.0) % 360.0))), drop=True)
+            fig = Figure(figsize=(long2-long1, lat2-lat1), facecolor='none', dpi=500)
+            ax = fig.add_axes([0, 0, 1, 1], projection=ccrs.PlateCarree(), frameon=False)
+            ax.set_extent([long1, long2, lat1, lat2], crs=ccrs.PlateCarree())
+            # ax.set_global()
+            # ax.coastlines()  
+            ax.barbs(data.longitude.values, data.latitude.values, data.u.values, data.v.values, 
+                    flagcolor='grey',
+                    # sizes=dict(emptybarb=0.25, spacing=0.2, height=0.5), linewidth=0.95, length=5, 
+                    transform=ccrs.PlateCarree())
+            buf = BytesIO()
+            fig.savefig(buf, format="png")
+            uri = "data:image/png;base64," + base64.b64encode(buf.getbuffer()).decode("ascii")
+            return [
+                {
+                    "id": "document",
+                    "name": "Weather",
+                    "version": "1.0",
+                },
+                {
+                    "id": "Weather",
+                    "rectangle": {
+                        "coordinates": {
+                            "wsenDegrees": [long1, lat1, long2, lat2],
+                        },
[…]
