+++ airtrafficsim/server/replay.py
@@ -4,12 +4,13 @@ import pandas as pd
 import numpy as np
 from datetime import datetime, timezone
 
+
 class Replay:
     @staticmethod
     def get_replay_dir():
         """
         Return a list of historic/simulation data directories in data/replay
-        
+
         Returns
         -------
         {}
@@ -19,9 +20,9 @@ class Replay:
         for dir in Path(__file__).parent.parent.parent.joinpath('data/flight_data').iterdir():
             if dir.is_dir():
                 historic_list.append(dir.name)
-        
-        simulation_list=[]
-        simulation_file_list={}
+
+        simulation_list = []
+        simulation_file_list = {}
         for dir in Path(__file__).parent.parent.parent.joinpath('result').iterdir():
             if dir.is_dir():
                 simulation_list.append(dir.name)
@@ -35,17 +36,16 @@ class Replay:
         simulation_list.sort(reverse=True)
         return {"historic": historic_list, "simulation": simulation_list, "simulation_files": simulation_file_list}
 
-
     @staticmethod
     def get_replay_czml(replayCategory, replayFile):
         """
         Generate CZML file for visualization given replay file name.
-        
+
         Parameters
         ----------
         replayCategory : string
             The category to replay (historic / simulation)
-        
+
         replayFile : string
             Name of the replay file directory
 
@@ -58,10 +58,10 @@ class Replay:
             trajectories = []
             start_time = None
             end_time = None
-            for file in Path(__file__).parent.parent.parent.joinpath('data/flight_data/',replayFile).iterdir():
+            for file in Path(__file__).parent.parent.parent.joinpath('data/flight_data/', replayFile).iterdir():
                 if file != Path(__file__).parent.parent.parent.joinpath('data/flight_data/', replayFile, replayFile+'.csv'):
                     file_content = pd.read_csv(file)
-                    
+
                     if replayCategory == 'historic':
                         # start = datetime.utcfromtimestamp(file_content.iloc[0]['timestamp'])
                         # end = datetime.utcfromtimestamp(file_content.iloc[-1]['timestamp'])
@@ -69,31 +69,35 @@ class Replay:
                         id = file.name
                         if ('timestamp' and 'long' and 'lat' and 'alt' and 'gspeed') in file_content:
                             # FR24
-                            start = datetime.fromtimestamp(file_content.iloc[0]['timestamp'], timezone.utc)
-                            end = datetime.fromtimestamp(file_content.iloc[-1]['timestamp'], timezone.utc)
-                            positions =  np.column_stack((file_content['timestamp'].map(lambda x : datetime.fromtimestamp(x, timezone.utc).isoformat()), 
-                                                    file_content['long'].values, file_content['lat'].values, file_content['alt'].values/3.2808)).flatten().tolist()
-                            label = [{"interval": datetime.fromtimestamp(time, timezone.utc).isoformat()+"/"+end.isoformat(), 
-                                    "string": file.name+"\n"+str(alt)+"ft "+str(gspeed)+"kt"} 
-                                    for time, alt, gspeed in zip(file_content['timestamp'], file_content['alt'], file_content['gspeed'])]
+                            start = datetime.fromtimestamp(
+                                file_content.iloc[0]['timestamp'], timezone.utc)
+                            end = datetime.fromtimestamp(
+                                file_content.iloc[-1]['timestamp'], timezone.utc)
+                            positions = np.column_stack((file_content['timestamp'].map(lambda x: datetime.fromtimestamp(x, timezone.utc).isoformat()),
+                                                         file_content['long'].values, file_content['lat'].values, file_content['alt'].values/3.2808)).flatten().tolist()
[…]
