+++ airtrafficsim/core/navigation.py
@@ -44,11 +44,13 @@ class Nav:
     # Install navigation data
     if not Path(__file__).parent.parent.parent.resolve().joinpath('./data/navigation/xplane/').is_dir():
         # Create directories
-        Path(__file__).parent.parent.parent.resolve().joinpath('./data/navigation/xplane/airports').mkdir(parents=True)
+        Path(__file__).parent.parent.parent.resolve().joinpath(
+            './data/navigation/xplane/airports').mkdir(parents=True)
 
         # Unzip files
         print("Unzipping X-plane navigation data.")
-        ZipFile(Path(__file__).parent.parent.parent.resolve().joinpath('./data/navigation/xplane_default_data.zip')).extractall(Path(__file__).parent.parent.parent.resolve().joinpath('./data/navigation/xplane/'))
+        ZipFile(Path(__file__).parent.parent.parent.resolve().joinpath('./data/navigation/xplane_default_data.zip')
+                ).extractall(Path(__file__).parent.parent.parent.resolve().joinpath('./data/navigation/xplane/'))
 
         # Extract apt.dat to runways.csv and individual csv in xplane/airports/
         print("Unpacking airport data (apt.dat). This will take a while...")
@@ -70,7 +72,8 @@ class Nav:
                     if row[0] in ("1", "16", "17", "99"):
                         # Write previous airport
                         if not icao == "":
-                            print("\r"+"Extracting information from airport "+icao, end="", flush=True)
+                            print(
+                                "\r"+"Extracting information from airport "+icao, end="", flush=True)
                             with open(Path(__file__).parent.parent.parent.resolve().joinpath('./data/navigation/xplane/airports', icao+'.csv'), 'w') as f:
                                 f.writelines(airport)
                         # Reset if not the end
@@ -103,19 +106,26 @@ class Nav:
 
     # Static variables
     print("Reading NAV data...")
-    fix = pd.read_csv(Path(__file__).parent.parent.parent.resolve().joinpath('./data/navigation/xplane/earth_fix.dat'), delimiter='\s+', skiprows=3, header=None)
+    fix = pd.read_csv(Path(__file__).parent.parent.parent.resolve().joinpath(
+        './data/navigation/xplane/earth_fix.dat'), delimiter='\s+', skiprows=3, header=None)
     """Fixes data https://developer.x-plane.com/wp-content/uploads/2019/01/XP-FIX1101-Spec.pdf"""
-    nav = pd.read_csv(Path(__file__).parent.parent.parent.resolve().joinpath('./data/navigation/xplane/earth_nav.dat'), delimiter='\s+', skiprows=3, header=None, names=np.arange(0,18), low_memory=False).apply(pd.to_numeric, errors='ignore')
+    nav = pd.read_csv(Path(__file__).parent.parent.parent.resolve().joinpath('./data/navigation/xplane/earth_nav.dat'),
+                      delimiter='\s+', skiprows=3, header=None, names=np.arange(0, 18), low_memory=False).apply(pd.to_numeric, errors='ignore')
     """Radio navigation data https://developer.x-plane.com/wp-content/uploads/2020/03/XP-NAV1150-Spec.pdf"""
[…]
