+++ paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 Air traffic management (ATM) research traditionally focuses on the macroscopic aspect of air transportation such as airspace design, traffic flow management, airport planning and scheduling, and more [@ATM]. Recently, with the new development of aerial vehicle concepts, including urban air mobility (UAM) and unmanned aircraft system (UAS), there has been a growing interest in performing microscopic ATM research, for example, conflict resolution using reinforcement learning [@Conflict], 4D-trajectory optimization [@4D], and even unmanned traffic management (UTM). Eurocontrol U-space [@uspace] and FAA/NASA UTM project [@nasa] are some examples of existing efforts in industry to perform such research. 
 
-To facilitate microscopic ATM research, an agent-based simulation and visualization software is needed. However, most ATM simulation tools are commercial products aimed at training air traffic controllers and airspace planning. ATM simulation tools for research purposes that are easily accessible and open-source, such as BlueSky [@BlueSky], are still scarce. Therefore, we developed AirTrafficSim to assist researchers to perform ATM research with an easy-to-use and comprehensive software environment to simulate air traffic movement and visualize the results. It is an open-source package that welcomes everyone to access and contribute.
+To facilitate microscopic ATM research, an agent-based simulation and visualization software is needed. However, most ATM simulation tools are commercial products aimed at training air traffic controllers and airspace planning. ATM simulation tools for research purposes that are easily accessible and open-source, such as BlueSky [@BlueSky], are still scarce. Therefore, we developed AirTrafficSim to assist researchers to perform ATM research with an easy-to-use and comprehensive software environment to simulate air traffic movement and visualize the results. Compare to existing solutions, AirTrafficSim provides weather intergration, a mordern UI, and an easy-to-user API to control the aircraft with algorithms such as reinforcement learning. It is an open-source package that welcomes everyone to access and contribute.
 
 # Summary
 
@@ -50,7 +50,7 @@ The autopilot module processes the assigned flight plan and controls the aircraf
 
 The performance module calculates the aircraft state, such as speed, heading, vertical rate, and fuel consumption, for each timestep. Currently, AirtrafficSim makes use of the licensed BADA performance model data from Eurocontrol [@bada] but it is extensible to other performance models such as the open-source OpenAP model [@openap].
 
-The flight route detection module detects the flight route including origin and destination airports, SIDs, and STARs from historical flight data and generates a flight plan for simulation. A robust algorithm is being developed to detect the airways to generate a complete flight plan.
+The flight route detection module detects the flight route including origin and destination airports, SIDs, and STARs from historical flight data and generates a flight plan for simulation.
 
 AirTrafficSim can be applied flexibly for different ATM research settings. One of the recent works is to simulate and validate the solutions to an arrival sequencing problem in the Hong Kong International Airport by applying a mixed-integer linear programming algorithm [@hkia]. The software can also be used to tackle conflict resolution problems, route coordination and optimization problems, contingency management problems, and more.
 
