### 2022-12-16 #4916 @harrylui1995
Hi @AoifeHughes, I am one of the co-authors of this paper, Go Nam Lui. I'm sorry for not getting back to you sooner. We read over @xoolive 's notes two weeks ago, and we plan to respond to him after taking care of all his comments. We will update you here in a few days.
