### 2023-01-17 HKUST-OCTAD-LAB/AirTrafficSim#32 @gmingas
Add tests
The codebase currently does not have any unit or integration tests. These should be added, at least for the most important methods. A continuous integration pipeline for testing could also be considered.

Part of JOSS review [here](https://github.com/openjournals/joss-reviews/issues/4916)
