### 2023-01-17 HKUST-OCTAD-LAB/AirTrafficSim#33 @gmingas
Simulation button in UI fails to load data
I installed locally and was able to see the UI and use it in replay mode. But when I click on the "Simulate" button in the bottom of the screen and select a scenario, it loads forever and does not show anything. Note that when I click replay and then select "Simulation" from the box that opens, I can run those scenarios.

This should be fixed or fail more gracefully if there is a reason (e.g. missing data files)

Part of JOSS review [here](https://github.com/openjournals/joss-reviews/issues/4916)
