### 2023-01-17 HKUST-OCTAD-LAB/AirTrafficSim#34 @gmingas
Clarify statement in paper
The following sentence was unclear to me while reading the article/paper submitted to JOSS:
"It can also control the aircraft based on the target state inputted by the user and algorithm to simulate ATC command". Could you elaborate on this in the paper or in the documentation?

Part of JOSS review [here](https://github.com/openjournals/joss-reviews/issues/4916)
