### 2023-01-17 HKUST-OCTAD-LAB/AirTrafficSim#35 @gmingas
Map sometimes fails to render
The map sometimes fails to fully render when I run replays. This also happens sometimes when I zoom in and out and when I switch satellite image on and off. Is there a way to make this smoother/more reliable?

See picture below for an example

Part of JOSS review [here](https://github.com/openjournals/joss-reviews/issues/4916)
