### 2023-01-17 HKUST-OCTAD-LAB/AirTrafficSim#36 @gmingas
Instructions on how to add BADA data
The installation instructions should go through the steps needed to add BADA data or some alternative. At the moment this is only discussed in the docs. As mentioned by the other reviewer I was able to run with BADA data because I have a license but other users might be confused and not be able to run the software.

Part of JOSS review [here](https://github.com/openjournals/joss-reviews/issues/4916)
