### 2023-01-17 HKUST-OCTAD-LAB/AirTrafficSim#37 @gmingas
Docstring completeness and format 
The docstrings format is somewhat inconsistent across the code. It would be good to stick to one format.

Also, some functions lack a doctring or have one which does not give much information. This could be improved to help contributions and future maintenance.

Part of JOSS review [here](https://github.com/openjournals/joss-reviews/issues/4916)
