### 2023-01-17 HKUST-OCTAD-LAB/AirTrafficSim#38 @gmingas
Add contribution guidelines in main folder
It would be good to add a contribution guide document at the top level of the repo, to summarise what process the users should follow when proposing changes, adding issues, code etc. Example [here](https://github.com/scikit-learn/scikit-learn/blob/main/CONTRIBUTING.md).

Part of JOSS review [here](https://github.com/openjournals/joss-reviews/issues/4916)
