### 2023-01-24 #4916 @kyfrankie
Dear Reviewers,

Thank you for your thorough comments. Your suggestions help a lot. Since we are all amateurs in managing open-source projects, the process is slow. We mainly draw @xoolive's comments here.

We have implemented a series of unit tests on the core functions and debugged some of the problems.

For the OpenAP implementation, we still plan to include it in the future, owing to the workload. We have updated documentation to specify the BADA version and the associated installation guidance and deleted the content related to OpenAP in the documentation and Readme.

As for building a python package, we found that because of the dependencies with the Cartopy package, Conda is better than pip. The current status is that we are still dealing with the conda build and the recipes. We will update you once it works.

Thanks for your kind suggestion for the data format. CSV format is first deployed to log aircraft movements in our software because of its row-major format. Besides, it facilitates monitoring purposes since it's human-readable. Later, we will implement the Parquet format to reduce storage consumption.

Regarding the ICAO format for the flight plan, the software only considers waypoints' ICAO name to build its plan currently. We plan to implement the function to read an ICAO format flight plan. We also updated the documentation for more clarity. The "simulation" and "replay" codes are explained in [Navigating the UI - AirTrafficSim (hkust-octad-lab.github.io)](https://hkust-octad-lab.github.io/AirTrafficSim/tutorial/UI.html). The algorithm to detect published airways is being developed and has yet to be available. The sentence has been removed to avoid confusion. Thank you for the suggestion.

 We also notice @gmingas' comments. We will update everyone again once the software has been updated to address the issues.

Lastly, happy Lunar New Year, everyone!

Best,
Frankie
