### 2023-01-26 #4916 @xoolive
@kyfrankie @harrylui1995 恭禧發財!

About pip vs conda, your package should describe the dependencies, regardless of pip and conda. If some of your dependencies require some external dependencies, you should leave the choice to the user to use one option or the other for the installation.

As a matter of fact, publishing a package on `conda-forge` requires that you have a package on pip as well...

Good luck anyway, don't hesitate to ping if you need help
