+++ airtrafficsim/__main__.py
@@ -1,6 +1,8 @@
 import sys
 import os
+from pathlib import Path
 from importlib import import_module
+from zipfile import ZipFile
 
 import airtrafficsim.server.server as server
 
@@ -11,6 +13,13 @@ def main():
         raise IOError(
             "BADA folder is empty. Remember to put the BADA performance data into /data/performance/BADA/.")
 
+    # Unpack client
+    if not Path(__file__).parent.parent.resolve().joinpath('./data/client/build/').is_dir():
+        print("Unzipping client.")
+        ZipFile(Path(__file__).parent.parent.resolve().joinpath('./data/client/build.zip')
+                ).extractall(Path(__file__).parent.parent.resolve().joinpath('./data/client/'))
+
+    # Handle command line arguments
     if len(sys.argv) > 1 and sys.argv[1] == '--headless':
         # Run user defined environment without UI: python -m airtrafficsim --headless <env name>
         Env = getattr(import_module('environment.' +
