+++ airtrafficsim/core/environment.py
@@ -39,7 +39,7 @@ class Environment:
         self.file_name = file_name+'-' + \
             self.datetime.isoformat(timespec='seconds')
         self.folder_path = Path(__file__).parent.parent.parent.resolve().joinpath(
-            'result/'+self.file_name)
+            'data/result/'+self.file_name)
         self.folder_path.mkdir()
         self.file_path = self.folder_path.joinpath(self.file_name+'.csv')
         self.writer = csv.writer(open(self.file_path, 'w+'))
@@ -174,7 +174,7 @@ class Environment:
                                          ].to_numpy().flatten().tolist()
                 label = [{"interval": time+"/"+(self.start_time + timedelta(seconds=self.end_time)).isoformat(),
                           "string": call_sign+"\n"+str(np.floor(alt))+"ft "+str(np.floor(cas))+"kt"}
-                         for time, alt, cas in zip(content.iloc[:, 2].to_numpy(), content.iloc[:, 5].to_numpy(dtype=np.float), content.iloc[:, 6].to_numpy(dtype=np.float))]
+                         for time, alt, cas in zip(content.iloc[:, 2].to_numpy(), content.iloc[:, 5].to_numpy(dtype=float), content.iloc[:, 6].to_numpy(dtype=float))]
 
                 trajectory = {
                     "id": call_sign,
