+++ .github/workflows/tests.yml
@@ -1,0 +1,52 @@
+name: Tests
+
+on:
+  push:
+    branches: [ "main" ]
+  pull_request:
+    branches: [ "main" ]
+
+jobs:
+  build-linux:
+    runs-on: ubuntu-latest
+    strategy:
+      max-parallel: 5
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python 3.11
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.11'
+    - name: Add conda to system path
+      run: |
+        # $CONDA is an environment variable pointing to the root of the miniconda directory
+        echo $CONDA/bin >> $GITHUB_PATH
+    - name: Install dependencies
+      run: |
+        conda env update --file environment.yml --name base
+    - name: Download data from private repository
+        uses: actions/checkout@v3
+        with:
+          repository: HKUST-OCTAD-LAB/AirTrafficSim-data
+          ssh-key: ${{ secrets.SSH_KEY }}
+    - name: Unzip data
+        run: |
+          ls
+          ls AirTrafficSim-data
+          unzip AirTrafficSim-data/BADA.zip 
+          mv AirTrafficSim-data/.cdsapirc $HOME/
+          ls
+        shell: bash
+    # - name: Lint with flake8
+    #   run: |
+    #     conda install flake8
+    #     # stop the build if there are Python syntax errors or undefined names
+    #     flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+    #     # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+    #     flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+    - name: Test with pytest
+      run: |
+        conda install pytest coverage
+        coverage run -m pytest
+        coverage report xml
