+++ README.md
@@ -4,7 +4,7 @@ AirTrafficSim is a web-based air traffic simulation software written in Python a
 
 ![Tests](https://github.com/HKUST-OCTAD-LAB/AirTrafficSim/actions/workflows/tests.yml/badge.svg)
 [![Code Coverage](https://img.shields.io/codecov/c/github/HKUST-OCTAD-LAB/AirTrafficSim.svg)](https://codecov.io/gh/HKUST-OCTAD-LAB/AirTrafficSim)
-[![Docs](https://github.com/HKUST-OCTAD-LAB/AirTrafficSim/actions/workflows/docs.yml/docs.svg)](https://hkust-octad-lab.github.io/AirTrafficSim/)
+[![Docs](https://github.com/HKUST-OCTAD-LAB/AirTrafficSim/actions/workflows/docs.yml/badge.svg)](https://hkust-octad-lab.github.io/AirTrafficSim/)
 [![status](https://joss.theoj.org/papers/7d4a9fdfae0c862863fa3645d3ae80b1/status.svg)](https://joss.theoj.org/papers/7d4a9fdfae0c862863fa3645d3ae80b1)
 <!-- ![Conda version](https://img.shields.io/conda/vn/conda-forge/) -->
 
