+++ .github/workflows/tests.yml
@@ -27,27 +27,51 @@ jobs:
 
     name: ${{ matrix.label }}
     runs-on: ${{ matrix.os }}
+    
+    defaults:
+      run:
+        shell: bash -el {0}
 
     steps:
     - uses: actions/checkout@v3
-
+    
+    - name: Cache conda
+      uses: actions/cache@v2
+      env:
+        # Increase this value to reset cache if etc/example-environment.yml has not changed
+        CACHE_NUMBER: 0
+      with:
+        path: ~/conda_pkgs_dir
+        key:
+          ${{ runner.os }}-conda-${{ env.CACHE_NUMBER }}-${{ hashFiles('environment.yml') }}
+          
     - name: Setup miniconda
       uses: conda-incubator/setup-miniconda@v2
       with:
         activate-environment: airtrafficsim
+        channel-priority: strict
         environment-file: environment.yml
         python-version: 3.11
-        auto-activate-base: false
+        use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
+          
+#     - name: Setup miniconda
+#       uses: conda-incubator/setup-miniconda@v2
+#       with:
+#         activate-environment: airtrafficsim
+#         environment-file: environment.yml
+#         python-version: 3.11
+#         auto-activate-base: false
+        
     - run: |
         conda info
-        conda list
+        conda list -n airtrafficsim
 
-    - name: Cache conda enviornment
-      uses: actions/cache@v2
-      with:
-        path: ${{ matrix.prefix }}
-        key: ${{ matrix.label }}-conda-${{ hashFiles('environment.yml') }}
-      id: cache
+#     - name: Cache conda enviornment
+#       uses: actions/cache@v2
+#       with:
+#         path: ${{ matrix.prefix }}
+#         key: ${{ matrix.label }}-conda-${{ hashFiles('environment.yml') }}
+#       id: cache
 
 #     - name: Update environment
 #       run: conda env update -n airtrafficsim -f environment.yml
@@ -73,7 +97,6 @@ jobs:
       run: |
         unzip AirTrafficSim-data/BADA.zip -d data/performance/BADA
         mv AirTrafficSim-data/.cdsapirc $HOME/
-      shell: bash
 
     # - name: Lint with flake8
     #   run: |
@@ -87,7 +110,6 @@ jobs:
       run: |
         coverage run -m pytest
         coverage report xml
-      shell: bash
 
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
