+++ .github/workflows/tests.yml
@@ -53,39 +53,11 @@ jobs:
         environment-file: environment.yml
         python-version: 3.11
         use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
-          
-#     - name: Setup miniconda
-#       uses: conda-incubator/setup-miniconda@v2
-#       with:
-#         activate-environment: airtrafficsim
-#         environment-file: environment.yml
-#         python-version: 3.11
-#         auto-activate-base: false
         
     - run: |
         conda info
-        conda list -n airtrafficsim
-
-#     - name: Cache conda enviornment
-#       uses: actions/cache@v2
-#       with:
-#         path: ${{ matrix.prefix }}
-#         key: ${{ matrix.label }}-conda-${{ hashFiles('environment.yml') }}
-#       id: cache
-
-#     - name: Update environment
-#       run: conda env update -n airtrafficsim -f environment.yml
-#       if: steps.cache.outputs.cache-hit != 'true'
-
-    # - name: Add conda to system path
-    #   run: |
-    #     # $CONDA is an environment variable pointing to the root of the miniconda directory
-    #     echo $CONDA/bin >> $GITHUB_PATH
-
-    # - name: Install dependencies
-    #   run: |
-    #     conda env update --file environment.yml --name base
-
+        conda list
+        
     - name: Download data from private repository
       uses: actions/checkout@v3
       with:
