+++ .github/workflows/tests.yml
@@ -79,6 +79,7 @@ jobs:
     #     flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
 
     - name: Test with pytest
+      continue-on-error: true
       run: |
         coverage run -m pytest
         coverage report xml
