+++ .github/workflows/tests.yml
@@ -82,7 +82,7 @@ jobs:
       continue-on-error: true
       run: |
         coverage run -m pytest
-        coverage report xml
+        coverage xml
 
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
