+++ .github/workflows/tests.yml
@@ -82,6 +82,8 @@ jobs:
       run: |
         coverage run -m pytest
         coverage xml
+        ls
+        ls /home/runner/work/AirTrafficSim/AirTrafficSim
 
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
