+++ .github/workflows/tests.yml
@@ -82,8 +82,6 @@ jobs:
       run: |
         coverage run -m pytest
         coverage xml
-        ls
-        ls /home/runner/work/AirTrafficSim/AirTrafficSim
 
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
