+++ paper.md
@@ -26,9 +26,9 @@ bibliography: paper.bib
 
 # Statement of need
 
-Air traffic management (ATM) research traditionally focuses on the macroscopic aspect of air transportation such as airspace design, traffic flow management, airport planning and scheduling, and more [@ATM]. Recently, with the new development of aerial vehicle concepts, including urban air mobility (UAM) and unmanned aircraft system (UAS), there has been a growing interest in performing microscopic ATM research, for example, conflict resolution using reinforcement learning [@Conflict], 4D-trajectory optimization [@4D], and even unmanned traffic management (UTM). Eurocontrol U-space [@uspace] and FAA/NASA UTM project [@nasa] are some examples of existing efforts in industry to perform such research. 
+Air traffic management (ATM) research traditionally focuses on the macroscopic aspect of air transportation such as airspace design, traffic flow management, airport planning and scheduling, and more [@ATM]. Recently, with the new development of aerial vehicle concepts, including urban air mobility (UAM) and unmanned aircraft system (UAS), there has been a growing interest in performing microscopic ATM research, for example, conflict resolution using reinforcement learning [@Conflict], 4D-trajectory optimization [@4D], and even unmanned traffic management (UTM). Eurocontrol U-space [@uspace] and FAA/NASA UTM project [@nasa] are some examples of existing efforts in the industry to perform such research. 
 
-To facilitate microscopic ATM research, an agent-based simulation and visualization software is needed. However, most ATM simulation tools are commercial products aimed at training air traffic controllers and airspace planning. ATM simulation tools for research purposes that are easily accessible and open-source, such as BlueSky [@BlueSky], are still scarce. Therefore, we developed AirTrafficSim to assist researchers to perform ATM research with an easy-to-use and comprehensive software environment to simulate air traffic movement and visualize the results. Compare to existing solutions, AirTrafficSim provides weather intergration, a mordern UI, and an easy-to-user API to control the aircraft with algorithms such as reinforcement learning. It is an open-source package that welcomes everyone to access and contribute.
+To facilitate microscopic ATM research, an agent-based simulation and visualization software is needed. However, most ATM simulation tools are commercial products aimed at training air traffic controllers and airspace planning. ATM simulation tools for research purposes that are easily accessible and open-source, such as BlueSky [@BlueSky], are still scarce. Therefore, we developed AirTrafficSim to assist researchers to perform ATM research with an easy-to-use and comprehensive software environment to simulate air traffic movement and visualize the results. Compare to existing solutions, AirTrafficSim provides weather integration, a modern UI, and an easy-to-use API to control the aircraft with algorithms such as reinforcement learning. It is an open-source package that welcomes everyone to access and contribute.
 
 # Summary
 
@@ -47,7 +47,7 @@ The navigation module provides global airports, waypoints, navigation aids and f
 
 The weather module provides weather information including multi-level wind, pressure, temperature, and single-level surface precipitation data from the ECMWF ERA5 weather database [@era5]. It also processes radar images provided by users as a source of high-resolution convective weather information.
 
-The autopilot module processes the assigned flight plan and controls the aircraft to follow the plan from take-off to landing. It can also control the aircraft based on the target state inputted by the user and algorithm to simulate ATC commands. Non-standard maneuvers that are sometimes used by air traffic controllers such as vectoring and holding can also be commanded.
+The autopilot module processes the assigned flight plan and controls the aircraft to follow the plan from take-off to landing. It can also control the movement of aircraft with functions simulating ATC commands, enabling traffic flow simulation with user-designed algorithms. Non-standard maneuvers that are sometimes used by air traffic controllers such as vectoring and holding can also be commanded.
 
[…]
