+++ .github/workflows/receivePR.yml
@@ -1,0 +1,19 @@
+name: Receive PR
+
+# read-only repo token
+# no access to secrets
+
+on:
+  pull_request:
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:        
+      - uses: actions/checkout@v3
+
+      - uses: actions/upload-artifact@v2
+        with:
+          name: pr
+          path: .
