+++ .github/workflows/tests.yml
@@ -33,10 +33,9 @@ jobs:
     name: ${{ matrix.label }}
     runs-on: ${{ matrix.os }}
 
-    if: >
-      github.repository == 'HKUST-OCTAD-LAB/AirTrafficSim' &&
-      github.event.workflow_run.conclusion == 'success'
+    if: github.event.workflow_run.conclusion == 'success'
     # if: contains(github.event.pull_request.labels.*.name, 'Ready for testing')
+    # if: github.repository == 'HKUST-OCTAD-LAB/AirTrafficSim'
     
     defaults:
       run:
@@ -67,6 +66,7 @@ jobs:
           fs.writeFileSync('${{github.workspace}}/pr.zip', Buffer.from(download.data));
     - run: |
         unzip pr.zip
+        unzip AirTrafficSim.zip -d .
         ls
     
     - name: Cache conda
