+++ .github/workflows/receivePR.yml
@@ -1,24 +1,0 @@
-name: Receive PR
-
-# read-only repo token
-# no access to secrets
-
-on:
-  pull_request:
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-
-    steps:        
-      - uses: actions/checkout@v3
-
-      - name: Zip folder
-        run: |
-          zip -r AirTrafficSim.zip .
-
-      - uses: actions/upload-artifact@v3
-        with:
-          name: pr
-          path: AirTrafficSim.zip
-          retention-days: 1
