+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "AirTrafficSim"
-version = "0.0.2"
+version = "0.0.3"
 authors = [
     {name = "Frankie Hui", email = "kyhuiaf@connect.ust.hk"},
 ]
@@ -37,7 +37,13 @@ repository = "https://github.com/HKUST-OCTAD-LAB/AirTrafficSim"
 airtrafficsim = "airtrafficsim.__main__:main"
 
 [tool.setuptools]
-packages = ["airtrafficsim", "airtrafficsim.core", "airtrafficsim.data", "airtrafficsim.server", "airtrafficsim.utils"]
+packages = ["airtrafficsim", 
+            "airtrafficsim.core",
+            "airtrafficsim.core.weather",
+            "airtrafficsim.core.performance",
+            "airtrafficsim.data", 
+            "airtrafficsim.server", 
+            "airtrafficsim.utils"]
 include-package-data = true
 
 [project.optional-dependencies]
