+++ README.md
@@ -6,17 +6,17 @@
 [![Conda version](https://img.shields.io/conda/vn/conda-forge/airtrafficsim)](https://anaconda.org/conda-forge/airtrafficsim)
 [![status](https://joss.theoj.org/papers/7d4a9fdfae0c862863fa3645d3ae80b1/status.svg)](https://joss.theoj.org/papers/7d4a9fdfae0c862863fa3645d3ae80b1)
 
-AirTrafficSim is a web-based air traffic simulation software written in Python and javascript. It is designed to visualize historical and research data, perform microscopic studies of air traffic movement with the integration of a historical weather database, and evaluate the performance of ATM algorithms.
+AirTrafficSim is a web-based air traffic simulation software written in Python and javascript. It is designed to visualize historical and simulated flight data and perform microscopic studies of air traffic movement with the integration of a historical weather database. It aims to assist users to evaluate the performance of ATM algorithms.
 
 ## Features
 
-- Replay histortic flights given data (FlightRadar 24 and simulated flights)
+- Replay historical flights with user-provided data (e.g. from FlightRadar 24)
 - Air traffic simulation using [BADA 3.15 performance data](https://www.eurocontrol.int/model/bada)
 - Navigation data simulation and visualization from [x-plane 11](https://developer.x-plane.com/docs/data-development-documentation/)
 - Autopilot and Flight Management System simulation
-- ATC traffic control methodology (Holding, vectoring, direct) simulation
-- Weather data from [ECMWF ERA5](https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-pressure-levels?tab=overview) and custom radar image
-- Air traffic is controlled with API interface to simulate ATC interaction
+- ATC commands (e.g. holding, vectoring, and direct to) simulation
+- Simulation and visualisation with weather data from [ECMWF ERA5](https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-pressure-levels?tab=overview) and custom radar image (HKO 256km radar images)
+- Aircraft is controlled with  an API interface to simulate ATC interaction
 
 ![AirTrafficSim](docs/source/images/UI_features.png)
 
@@ -37,7 +37,7 @@ If you find AirTrafficSim useful for your research, please cite the following un
 
 ## Installation
 
-The latest stable release of AirTrafficSim can be installed from conda-forge. All dependencies will be installated automatically with the client pre-built and ready for use directly after installation.
+Linux or [WSL on window](https://learn.microsoft.com/en-us/windows/wsl/) is recommended for AirTrafficSim. The latest stable release of AirTrafficSim can be installed from conda-forge. All dependencies will be installed automatically with the web client pre-built and ready for use directly after installation.
 
 It is recommended to install AirTrafficSim in a new conda environment:
 
@@ -45,27 +45,27 @@ It is recommended to install AirTrafficSim in a new conda environment:
 conda create -n airtrafficsim -c conda-forge airtrafficsim 
 ```
 
-Then, please initialise AirTrafficSim by specifing a folder path to create a symbolic link to the [airtrafficsim_data](airtrafficsim/data/) folder:
+Then, please initialise AirTrafficSim by specifying a folder path to create a symbolic link to the [airtrafficsim_data](airtrafficsim/data/) folder:
 
 ```
 conda activate airtrafficsim
 airtrafficsim -- init <path to a folder>
 ```
-You may provide or retrieve any data of AirTrafficSim and create simulation environments in this folder. Please visit the [documentation](https://hkust-octad-lab.github.io/AirTrafficSim/index.html) for more information.
[…]
