+++ .github/workflows/docs.yml
@@ -37,7 +37,8 @@ jobs:
         use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
         
     - name: Build HTML
-      run: sphinx-build -b html docs/source/ docs/build/html
+      run: |
+        sphinx-build -b html docs/source/ docs/build/html
         
     - name: Deploy
       uses: peaceiris/actions-gh-pages@v3
