### 2023-05-02 #4916 @kyfrankie
Dear editor and reviewers,

Thank you a lot for your patience and all your comments. They are very useful and have helped us to develop a better open-source software package. We have improved AirTrafficSim to address the comments and suggestions and I would like to provide a summary below:

- Completed set up of automated testing with pytest and Coverage.py. Integration testing has been implemented.
- Packaged and published AirTrafficSim on the conda-forge channel. Related installation instruction has been updated.
- Initial integration of OpenAP open-source aircraft performance model.
- Fixed and improved code quality, dead imports and docstrings.
- Clarified and improved paper and documentation wording.
- Added contribution guideline.
- Specified BADA version and updated installation guide.
- Updated UI to address [issue](https://github.com/HKUST-OCTAD-LAB/AirTrafficSim/issues/33)

We hope this can address the issues and we would like to express our sincere gratitude for your patience and time to edit and review this work. We will continue to work on developing AirTrafficSim to improve its code quality and features, for example, using better file structure than CSV files, using ICAO flight plan format, using the [Traffic](https://traffic-viz.github.io/) package to handle OpenSky data, improving UI, adding more test cases, and more.

Please feel free to let me know if you have any questions. Thanks.
