### 2023-05-24 #4916 @xoolive
Hi @kyfrankie @harrylui1995 @AoifeHughes

Very good job!! There has been significant improvement since last time.

Few remarks though.

About the paper:
- line 11: I would recommend that you don't use the qualifier "microscopic" for ATM research. There's not much that is microscopic in ATM...
- line 22: typo: "compare **d** to"
- line 23: maybe "control the aircraft with external modules" rather than limiting yourself to "reinforcement learning"
- line 42: can we use other sources of (possibly closed) sources of navigation data
- general comment: why not mentioning the ability to replay existing traffic (from ADS-B for instance)

In the documentation:
- fix `-- init` to `--init`

There are still quite some glitches in the code, but nothing critical to me.
I will tick all the boxes, so I am not blocking the process, but please consider these comments.

Please keep in touch for integration with [traffic](https://github.com/xoolive/traffic)
