### 2023-05-25 #4916 @gmingas
@kyfrankie @harrylui1995 Thanks so much for the changes, it looks really food now and congrats for packaging the library and putting it in conda.

I updated my checklist and the only things left to do I think are a couple of changes to the documentation:
- Make sure docs point to BADA website where users can get a license, to avoid them having to search for it.
- Change the `-- init` to `--init`, as it currently fails (mentioned above by @xoolive)
