+++ paper.md
@@ -26,9 +26,9 @@ bibliography: paper.bib
 
 # Statement of need
 
-Air traffic management (ATM) research traditionally focuses on the macroscopic aspect of air transportation such as airspace design, traffic flow management, airport planning and scheduling, and more [@ATM]. Recently, with the new development of aerial vehicle concepts, including urban air mobility (UAM) and unmanned aircraft system (UAS), there has been a growing interest in performing microscopic ATM research, for example, conflict resolution using reinforcement learning [@Conflict], 4D-trajectory optimization [@4D], and even unmanned traffic management (UTM). Eurocontrol U-space [@uspace] and FAA/NASA UTM project [@nasa] are some examples of existing efforts in the industry to perform such research. 
+Air traffic management (ATM) research traditionally focuses on the macroscopic aspect of air transportation such as airspace design, traffic flow management, airport planning and scheduling, and more [@ATM]. Recently, with the new development of aerial vehicle concepts, including urban air mobility (UAM) and unmanned aircraft system (UAS), there has been a growing interest in performing ATM research, for example, conflict resolution using reinforcement learning [@Conflict], 4D-trajectory optimization [@4D], and even unmanned traffic management (UTM). Eurocontrol U-space [@uspace] and FAA/NASA UTM project [@nasa] are some examples of existing efforts in the industry to perform such research. 
 
-To facilitate microscopic ATM research, an agent-based simulation and visualization software is needed. However, most ATM simulation tools are commercial products aimed at training air traffic controllers and airspace planning. ATM simulation tools for research purposes that are easily accessible and open-source, such as BlueSky [@BlueSky], are still scarce. Therefore, we developed AirTrafficSim to assist researchers to perform ATM research with an easy-to-use and comprehensive software environment to simulate air traffic movement and visualize the results. Compare to existing solutions, AirTrafficSim provides weather integration, a modern UI, and an easy-to-use API to control the aircraft with algorithms such as reinforcement learning. It is an open-source package that welcomes everyone to access and contribute.
+To facilitate microscopic ATM research, an agent-based simulation and visualization software is needed. However, most ATM simulation tools are commercial products aimed at training air traffic controllers and airspace planning. ATM simulation tools for research purposes that are easily accessible and open-source, such as BlueSky [@BlueSky], are still scarce. Therefore, we developed AirTrafficSim to assist researchers to perform ATM research with an easy-to-use and comprehensive software environment to simulate air traffic movement and visualize the results. Compared to existing solutions, AirTrafficSim provides weather integration, a modern UI, and an easy-to-use API to control the aircraft with external modules. It is an open-source package that welcomes everyone to access and contribute.
 
 # Summary
 
@@ -37,7 +37,7 @@ To facilitate microscopic ATM research, an agent-based simulation and visualizat
 
 AirTrafficSim is a web-based air traffic simulation software written in Python and JavaScript. It is designed to visualize historical and research data, perform microscopic studies of air traffic movement with the integration of a weather database, and evaluate the performance of ATM algorithms. \autoref{fig:Architecture} shows the architecture of AirTrafficSim.
 
-AirTrafficSim contains a web-based frontend written in JavaScript using the Ionic React framework to provide an easy-to-use user interface (UI) to visualize both historical and simulated air traffic and other data in a browser. The 3D modeling of the globe is supported by the CesiumJS library to stream high-resolution maps, terrains, and 3D building data. The library also provides a rich API to visualize dynamic geospatial data obtained from performing simulations. The UI can also plot aircraft parameters using the Plotly.js library. 
[…]
