### 2023-05-26 94f66a6
Update paper.md
Add corresponding author to Go Nam Lui
