### 2023-06-19 #4916 @AoifeHughes
Morning @harrylui1995, Just a few quick thoughts on the `paper.md`

- "timestep" -> "time step"
- "an easy-to-use API..." this sentence could be reworked, I wouldn't say that "reinforcement learning" is an algorithm. So maybe even just dropping the "such as" part of the sentence.
- Figure 1, could you add more text to the legend so that the figure can be more easily understood
- Add a citation for "Iconic react"
    - and Flask + other key libraries you've mentioned.
