### 2023-06-19 d2aa0c8
Update paper.bib
Update some bibs based on editor's comment
