+++ paper.bib
@@ -117,3 +117,70 @@
     title  = {Tactical Routing for Air Transportation in HKIA Terminal Manuevering Area},
     booktitle={The 26th HKSTS International Conference}
 }
+@article{lui2022weather,
+  title={Weather impact quantification on airport arrival on-time performance through a Bayesian statistics modeling approach},
+  author={Lui, Go Nam and Hon, Kai Kwong and Liem, Rhea P},
+  journal={Transportation Research Part C: Emerging Technologies},
+  volume={143},
+  pages={103811},
+  year={2022},
+  publisher={Elsevier}
+}
+@inproceedings{lui2020towards,
+  title={Towards understanding the impact of convective weather on aircraft arrival traffic at the Hong Kong International Airport},
+  author={Lui, Go Nam and Liem, Rhea Patricia and Hon, Kai Kwong},
+  booktitle={IOP Conference Series: Earth and Environmental Science},
+  volume={569},
+  number={1},
+  pages={012067},
+  year={2020},
+  organization={IOP Publishing}
+}
+@misc{ionic,
+  title = {Ionic},
+  year = {2021},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/ionic-team/ionic-framework}},
+  note = {Accessed September 1, 2021},
+}
+@inproceedings{schafer2014bringing,
+  title={Bringing up {OpenSky: A large-scale ADS-B }sensor network for research},
+  author={Sch{\"a}fer, Matthias and Strohmeier, Martin and Lenders, Vincent and Martinovic, Ivan and Wilhelm, Matthias},
+  booktitle={Proceedings of the $13^{\text{th}}$ International Symposium on Information Processing in Sensor Networks},
+  pages={83--94},
+  year={2014},
+  organization={IEEE Press}
+}
+
+@Misc{Flightradar24,
+howpublished = {\url{https://www.flightradar24.com/data.com}},
+note = {Accessed July 1, 2019},
+title = {Data},
+author = {Flightradar24},
+year ={2019}
+}
+@Misc{celsium,
+howpublished = {\url{https://cesium.com/platform/cesiumjs/}},
+note = {Accessed September 1, 2021},
+title = {$3$D geospatial visualization for the web},
+author = {CesiumJS},
+year ={2021}
+}
+
+@misc{plotlyjs,
+  title = {Open-source JavaScript charting library behind Plotly and Dash},
+  author = {Plotly.js},
+  year = {2021},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/plotly/plotly.js/}},
+  note = {Accessed September 1, 2021},
+}
+@Misc{flask,
+howpublished = {\url{https://flask.palletsprojects.com/}},
+note = {Accessed September 1, 2021},
+title = {Flask},
+year ={2021}
+}
+
