+++ paper.bib
@@ -117,6 +117,7 @@
     title  = {Tactical Routing for Air Transportation in HKIA Terminal Manuevering Area},
     booktitle={The 26th HKSTS International Conference}
 }
+
 @article{lui2022weather,
   title={Weather impact quantification on airport arrival on-time performance through a Bayesian statistics modeling approach},
   author={Lui, Go Nam and Hon, Kai Kwong and Liem, Rhea P},
@@ -126,6 +127,7 @@
   year={2022},
   publisher={Elsevier}
 }
+
 @inproceedings{lui2020towards,
   title={Towards understanding the impact of convective weather on aircraft arrival traffic at the Hong Kong International Airport},
   author={Lui, Go Nam and Liem, Rhea Patricia and Hon, Kai Kwong},
@@ -136,14 +138,16 @@
   year={2020},
   organization={IOP Publishing}
 }
+
 @misc{ionic,
-  title = {Ionic},
-  year = {2021},
+  title = {Ionic framework},
+  author = {Ionic},
+  year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/ionic-team/ionic-framework}},
-  note = {Accessed September 1, 2021},
 }
+
 @inproceedings{schafer2014bringing,
   title={Bringing up {OpenSky: A large-scale ADS-B }sensor network for research},
   author={Sch{\"a}fer, Matthias and Strohmeier, Martin and Lenders, Vincent and Martinovic, Ivan and Wilhelm, Matthias},
@@ -155,32 +159,31 @@
 
 @Misc{Flightradar24,
 howpublished = {\url{https://www.flightradar24.com/data.com}},
-note = {Accessed July 1, 2019},
-title = {Data},
+title = {Flightradar24 ADS-B data},
 author = {Flightradar24},
-year ={2019}
+year ={2013}
 }
+
 @Misc{celsium,
 howpublished = {\url{https://cesium.com/platform/cesiumjs/}},
-note = {Accessed September 1, 2021},
 title = {$3$D geospatial visualization for the web},
 author = {CesiumJS},
-year ={2021}
+year ={2023}
 }
 
 @misc{plotlyjs,
   title = {Open-source JavaScript charting library behind Plotly and Dash},
   author = {Plotly.js},
-  year = {2021},
+  year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/plotly/plotly.js/}},
-  note = {Accessed September 1, 2021},
 }
+
 @Misc{flask,
 howpublished = {\url{https://flask.palletsprojects.com/}},
-note = {Accessed September 1, 2021},
 title = {Flask},
-year ={2021}
+author = {The Pallets Projects},
+year = {2023}
 }
 
