+++ paper.bib
@@ -161,7 +161,7 @@
 howpublished = {\url{https://www.flightradar24.com/data.com}},
 title = {Flightradar24 ADS-B data},
 author = {Flightradar24},
-year ={2013}
+year ={2023}
 }
 
 @Misc{celsium,
@@ -183,7 +183,7 @@ year ={2023}
 @Misc{flask,
 howpublished = {\url{https://flask.palletsprojects.com/}},
 title = {Flask},
-author = {The Pallets Projects},
+author = {PalletsProjects},
 year = {2023}
 }
 
