### 2023-06-22 #4916 @kyfrankie
Hi @AoifeHughes, thanks for the comments.

The figures and citations should have been updated in the paper already. Please let me know if further adjustment is needed. In addition, the software was tagged with release 0.1.0 and the archive was uploaded to Zenodo with doi [10.5281/zenodo.8070119](https://doi.org/10.5281/zenodo.8070119).

Thank you for much for the guidance along the process.
