+++ paper.bib
@@ -125,7 +125,8 @@
   volume={143},
   pages={103811},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.trc.2022.103811}
 }
 
 @inproceedings{lui2020towards,
@@ -136,7 +137,8 @@
   number={1},
   pages={012067},
   year={2020},
-  organization={IOP Publishing}
+  organization={IOP Publishing},
+  doi = {10.1088/1755-1315/569/1/012067}
 }
 
 @misc{ionic,
@@ -154,7 +156,8 @@
   booktitle={Proceedings of the $13^{\text{th}}$ International Symposium on Information Processing in Sensor Networks},
   pages={83--94},
   year={2014},
-  organization={IEEE Press}
+  organization={IEEE Press},
+  doi={10.1109/ipsn.2014.6846743}
 }
 
 @Misc{Flightradar24,
